pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["LuosRemmus <17summersoul17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pymystem3 = "^0.2.0"
pydantic = "^2.4.2"
aiohttp = "^3.8.6"
telethon = "^1.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
